WIP refactoring snapshotmap structure
