+++ vue/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@intlify/vue-i18n-loader": "^0.6.1",
     "@mapbox/geo-viewport": "^0.4.0",
+    "@maplibre/maplibre-gl-geocoder": "^1.3.1",
     "@vue/composition-api": "^1.0.0-rc.2",
     "@vue/test-utils": "^1.0.2",
     "apollo-cache-inmemory": "^1.6.5",
@@ -35,6 +36,7 @@
     "is-svg": "^4.2.2",
     "lodash": "^4.17.20",
     "mapbox.js": "^3.2.1",
+    "maplibre-gl": "^2.1.9",
     "normalize-url": "^4.5.1",
     "postcss": "^7.0.36",
     "ssri": "^7.1.1",
